requirements.txt
spotipy
python-dotenv
soundcloud-v2
rich
pytube
google-api-python-client
google-auth-oauthlib
google-auth-httplib2